优化了天勤行情数据服务

+++ CreateBackTestProcess_Inheritance.py
@@ -34,8 +34,9 @@ class BackTestWindow(QWidget, UI):  # 创建回测窗口类
         # 设置透明度
         self.setWindowOpacity(0.95)
 
-        self.parent.ioModal = self.parent.ioModal   # 创建读写csv文件对象
+        # self.parent.ioModule = self.parent.ioModule   # 创建读写csv文件对象
         self.pid = None
+        self.backtest_name = None
         self.add_paramer_to_combobox()
         self.label_info.setText("回测进程名的定义为：回测+当前时间+策略模板名+合约名+周期（分钟）\n创建回测进程时,策略模板名,合约名,周期三者不可缺少")
 
@@ -45,6 +46,9 @@ class BackTestWindow(QWidget, UI):  # 创建回测窗口类
         self.Btn_open_with_chrome.clicked.connect(self.open_backtest_web_with_sys_browser)          # 打开回测网页按钮点击事件槽函数
         self.calendar_backtest_startdate.selectionChanged.connect(self.get_backtest_date)           # 回测开始日期日历点击事件槽函数
         self.calendar_backtest_enddate.selectionChanged.connect(self.get_backtest_date)             # 回测结束日期日历点击事件槽函数
+
+        self.comboBox_exchange.activated.connect(self.add_contracts_to_comboBox_symbol)  # comboBox 信号槽函数
+        self.comboBox_contract_type.activated.connect(self.add_contracts_to_comboBox_symbol)
 
     def mousePressEvent(self, e):  # 鼠标点击事件
         if e.button() == Qt.LeftButton:
@@ -77,7 +81,7 @@ class BackTestWindow(QWidget, UI):  # 创建回测窗口类
 
     def get_tq_account_list(self):  # 获取天勤账号列表
         tq_account_list = []
-        data = self.parent.ioModal.read_csv_file(path='./data/tq_account.csv')
+        data = self.parent.ioModule.read_csv_file(path='./data/tq_account.csv')
         if data.empty:
             pass
         else:
@@ -119,102 +123,127 @@ class BackTestWindow(QWidget, UI):  # 创建回测窗口类
                                         pass                                    
         return class_name_list
 
+    def add_contracts_to_comboBox_symbol(self):
+        if self.comboBox_contract_type.currentText():
+            exchange = self.comboBox_exchange.currentText().split()[-1]
+            type = self.comboBox_contract_type.currentText().split()[-1]
+            pth = './available_contracts/' + exchange + '_' + type + '.csv'
+            if os.path.exists(pth):
+                data = self.parent.ioModule.read_csv_file(path=pth)
+                list = data['0'].tolist()
+                self.comboBox_symbol.clear()
+                self.comboBox_symbol.addItems(list)
+
     def get_backtest_parameters(self):  # 获取创建一个策略回测进程所需参数
-        df = self.parent.ioModal.read_csv_file(path='./data/tq_account.csv')
-
-        exchange = self.comboBox_exchange.currentText().split()[-1]  # 获取列表框选择的字符串分割后的最后一部分
-        symbol = exchange + '.' + self.symbol.text()
+        df = self.parent.ioModule.read_csv_file(path='./data/tq_account.csv')
         my_dict = {}
-
         now = datetime.now()
-        now_str = now.strftime('%Y.%m.%d-%H时%M分')                    # 获取当前时间,用于生成回测进程名
-        if self.comboBox_select_tq_account.currentText():
-            if self.comboBox_select_strategy.currentText():
-                if self.symbol.text():
-                    if self.symbol_period.text():
+        now_str = now.strftime('%Y.%m.%d-%H时%M分')                      # 获取当前时间,用于生成回测进程名
 
-                        # 回测进程名的定义为：回测+当前时间+策略名+合约名+周期（分钟），
-                        # 与正式运行的策略名区分开来，
-                        # 因为保存日志的时候是按进程名新建文件夹保存的，区分开来便于查找
-                        process_name = (
-                            '回测' + '-{' + now_str +'}-'
-                            + self.comboBox_select_strategy.currentText() + '-'
-                            + symbol + '-' + self.symbol_period.text() + 'min'
-                        )
+        if self.comboBox_select_tq_account.currentText():               # 检查是否有天勤账户
+            if self.comboBox_select_strategy.currentText():             # 检查是否有策略模板
+                if self.comboBox_contract_type.currentText():           # 检查合约类型栏是否为空
+                    if self.comboBox_symbol.currentText():              # 检查合约栏是否有内容
+                        path_contract = ('./available_contracts/' + self.comboBox_exchange.currentText().split()[-1] + '_' + self.comboBox_contract_type.currentText().split()[-1] + '.csv')
 
-                        my_dict['process_name'] = process_name  # 进程名称或策略实例名称
-                        my_dict['client_name'] = ' '  # 回测时不需要客户名称
-                        my_dict['tq_account'] = self.comboBox_select_tq_account.currentText()  # 天勤账号
+                        if os.path.exists(path_contract):  # 检查是否有合约表文件
+                            data = self.parent.ioModule.read_csv_file(path=path_contract)
+                            list = data['0'].tolist()
 
-                        current_tq_account = self.comboBox_select_tq_account.currentText()
-                        index = df.index[df['tq_account'] == current_tq_account]  # 获取当前所选项目对应的pd行index
-                        for idx, row in df.iterrows():
-                            if idx == index:
-                                my_dict['tq_psd'] = str(row['tq_psd'])  # 天勤密码
-                                my_dict['futures_company'] = row['futures_company']  # 期货公司
-                                my_dict['futures_account'] = row['futures_account']  # 期货账号
-                                my_dict['futures_psd'] = row['futures_psd']          # 期货密码
-                                break
-                        my_dict['symbol'] = symbol  # 合约代码
-                        my_dict['symbol_period'] = self.symbol_period.text()                # 合约周期
-                        my_dict['strategy'] = self.comboBox_select_strategy.currentText()   # 策略名称
-                        my_dict['whether_self_start'] = True                              # 是否自启动，回测时此项默认为True
-                        my_dict['whether_live_trading'] = False                           # 是否实盘交易，回测时此项默认为False
-                        my_dict['whether_backtest'] = True                                # 是否回测
-                        my_dict['whether_open_web_services'] = True                       # 是否启动web服务，回测时此项默认为True
-                        my_dict['web_port'] = self.web_port.text()                          # 网页端口
+                            if self.comboBox_symbol.currentText() in list:  # 检查输入的合约在不在合约表文件里面
+                                if self.symbol_period.text():  # 检查是否有合约周期
 
-                        my_dict['trading_status'] = True                                    # 停止交易标志位，0为正常交易，1为停止交易
-                        my_dict['orientation'] = self.orientation.text()                    # 交易方向,用于半自动化策略,0为无方向,1或正整数为做多,-1或负整数为做空
-                        my_dict['initial_capital'] = self.initial_capital.text()            # 初始资金
-                        my_dict['final_capital'] = self.initial_capital.text()              # 最终资金
-                        my_dict['contract_multiples'] = self.contract_multiples.text()      # 合约倍数
-                        my_dict['margin_rate'] = self.margin_rate.text()                    # 保证金率
-                        my_dict['stop_loss'] = self.stop_loss.text()                        # 止损点
-                        my_dict['stop_profit'] = self.stop_profit.text()                    # 止盈点
+                                    # 回测进程名的定义为：回测+当前时间+策略名+合约名+周期（分钟），
+                                    # 与正式运行的策略名区分开来，
+                                    # 因为保存日志的时候是按进程名新建文件夹保存的，区分开来便于查找
+                                    process_name = (
+                                        '回测' + '-{' + now_str +'}-' + self.comboBox_select_strategy.currentText() + '-'
+                                        + self.comboBox_symbol.currentText() + '-' + self.symbol_period.text() + 'min'
+                                    )
 
-                        my_dict['long_add_times'] = 0  # 做多累积次数
-                        my_dict['long_current_position'] = 0  # 当前多单持仓
-                        my_dict['first_long_price'] = 0  # 第一次做多价格
-                        my_dict['first_long_deal'] = 0  # 第一次做多成交量
-                        my_dict['short_add_times'] = 0  # 做空累积次数
-                        my_dict['short_current_position'] = 0  # 当前空单持仓
-                        my_dict['first_short_price'] = 0  # 第一次做空价格
-                        my_dict['first_short_deal'] = 0  # 第一次做空成交量
+                                    my_dict['process_name'] = process_name  # 进程名称或策略实例名称
+                                    my_dict['client_name'] = ' '  # 回测时不需要客户名称
+                                    my_dict['tq_account'] = self.comboBox_select_tq_account.currentText()  # 天勤账号
 
-                        my_dict['whether_open_line'] = self.checkBox_whether_open_line.isChecked()  # 是否定义了开仓直线
-                        my_dict['open_line_Coordinates'] = self.open_line_Coordinates.text()  # 开仓线坐标
-                        my_dict['whether_close_line'] = self.checkBox_whether_close_line.isChecked()  # 是否定义了平仓直线
-                        my_dict['close_line_Coordinates'] = self.close_line_Coordinates.text()  # 平仓线坐标
+                                    current_tq_account = self.comboBox_select_tq_account.currentText()
+                                    index = df.index[df['tq_account'] == current_tq_account]  # 获取当前所选项目对应的pd行index
+                                    for idx, row in df.iterrows():
+                                        if idx == index:
+                                            my_dict['tq_psd'] = str(row['tq_psd'])  # 天勤密码
+                                            my_dict['futures_company'] = row['futures_company']  # 期货公司
+                                            my_dict['futures_account'] = row['futures_account']  # 期货账号
+                                            my_dict['futures_psd'] = row['futures_psd']          # 期货密码
+                                            break
+                                    my_dict['symbol'] = self.comboBox_symbol.currentText()  # 合约代码
+                                    my_dict['symbol_period'] = self.symbol_period.text()                # 合约周期
+                                    my_dict['strategy'] = self.comboBox_select_strategy.currentText()   # 策略名称
+                                    my_dict['whether_self_start'] = True                              # 是否自启动，回测时此项默认为True
+                                    my_dict['whether_live_trading'] = False                           # 是否实盘交易，回测时此项默认为False
+                                    my_dict['whether_backtest'] = True                                # 是否回测
+                                    my_dict['whether_open_web_services'] = True                       # 是否启动web服务，回测时此项默认为True
+                                    my_dict['web_port'] = self.web_port.text()                          # 网页端口
 
-                        my_dict['CP1'] = self.Customized_parameters1.text()         # 自定义参数1  Customized_parameters 为了方便使用,缩写为CP
-                        my_dict['CP2'] = self.Customized_parameters2.text()         # 自定义参数2
-                        my_dict['CP3'] = self.Customized_parameters3.text()         # 自定义参数3
-                        my_dict['CP4'] = self.Customized_parameters4.text()         # 自定义参数4
-                        my_dict['CP5'] = self.Customized_parameters5.text()         # 自定义参数5
-                        my_dict['CP6'] = self.Customized_parameters6.text()         # 自定义参数6
-                        my_dict['CP7'] = self.Customized_parameters7.text()         # 自定义参数7
-                        my_dict['CP8'] = self.Customized_parameters8.text()         # 自定义参数8
+                                    my_dict['trading_status'] = True                                    # 停止交易标志位，0为正常交易，1为停止交易
+                                    my_dict['orientation'] = self.orientation.text()                    # 交易方向,用于半自动化策略,0为无方向,1或正整数为做多,-1或负整数为做空
+                                    my_dict['initial_capital'] = self.initial_capital.text()            # 初始资金
+                                    my_dict['final_capital'] = self.initial_capital.text()              # 最终资金
+                                    my_dict['contract_multiples'] = self.contract_multiples.text()      # 合约倍数
+                                    my_dict['margin_rate'] = self.margin_rate.text()                    # 保证金率
+                                    my_dict['stop_loss'] = self.stop_loss.text()                        # 止损点
+                                    my_dict['stop_profit'] = self.stop_profit.text()                    # 止盈点
 
-                        index = 0
+                                    my_dict['long_add_times'] = 0  # 做多累积次数
+                                    my_dict['long_current_position'] = 0  # 当前多单持仓
+                                    my_dict['first_long_price'] = 0  # 第一次做多价格
+                                    my_dict['first_long_deal'] = 0  # 第一次做多成交量
+                                    my_dict['short_add_times'] = 0  # 做空累积次数
+                                    my_dict['short_current_position'] = 0  # 当前空单持仓
+                                    my_dict['first_short_price'] = 0  # 第一次做空价格
+                                    my_dict['first_short_deal'] = 0  # 第一次做空成交量
 
-                        if (self.label_backtest_startdate.text() == '') or (self.label_backtest_enddate.text() == ''):
-                            self.label_info.setText('请先选择回测开始日期和结束日期')
+                                    my_dict['whether_open_line'] = self.checkBox_whether_open_line.isChecked()  # 是否定义了开仓直线
+                                    my_dict['open_line_Coordinates'] = self.open_line_Coordinates.text()  # 开仓线坐标
+                                    my_dict['whether_close_line'] = self.checkBox_whether_close_line.isChecked()  # 是否定义了平仓直线
+                                    my_dict['close_line_Coordinates'] = self.close_line_Coordinates.text()  # 平仓线坐标
+
+                                    my_dict['CP1'] = self.Customized_parameters1.text()         # 自定义参数1  Customized_parameters 为了方便使用,缩写为CP
+                                    my_dict['CP2'] = self.Customized_parameters2.text()         # 自定义参数2
+                                    my_dict['CP3'] = self.Customized_parameters3.text()         # 自定义参数3
+                                    my_dict['CP4'] = self.Customized_parameters4.text()         # 自定义参数4
+                                    my_dict['CP5'] = self.Customized_parameters5.text()         # 自定义参数5
+                                    my_dict['CP6'] = self.Customized_parameters6.text()         # 自定义参数6
+                                    my_dict['CP7'] = self.Customized_parameters7.text()         # 自定义参数7
+                                    my_dict['CP8'] = self.Customized_parameters8.text()         # 自定义参数8
+                                    my_dict['CP9'] = self.Customized_parameters9.text()         # 自定义参数9
+
+                                    index = 0
+
+                                    if (self.label_backtest_startdate.text() == '') or (self.label_backtest_enddate.text() == ''):
+                                        self.label_info.setText('请先选择回测开始日期和结束日期')
+                                    else:
+                                        backtest_start_date = self.label_backtest_startdate.text()
+                                        backtest_end_date = self.label_backtest_enddate.text()
+                                        #定义一个元组，用于存储日期和字典
+                                        backtest_tuple = (index, my_dict, backtest_start_date, backtest_end_date)
+                                        print('传入回测的元组为：' + str(backtest_tuple))
+                                        self.start_backtest_process(backtest_tuple)
+
+
+                                else:
+                                    self.label_info.setText('请先输入回测周期,该项参数不可缺少')
+                            else:
+                                self.label_info.setText('输入的合约名不在天勤的可用合约列表里,请仔细检查')
                         else:
-                            backtest_start_date = self.label_backtest_startdate.text()
-                            backtest_end_date = self.label_backtest_enddate.text()
-                            #定义一个元组，用于存储日期和字典
-                            backtest_tuple = (index, my_dict, backtest_start_date, backtest_end_date)
-                            print('传入回测的元组为：' + str(backtest_tuple))
-                            self.start_backtest_process(backtest_tuple)
+                            self.label_info.setText('没有可用的合约表,请先运行一次天勤行情服务,下载可用合约表')
                     else:
-                        self.label_info.setText('请先输入回测周期,该项参数不可缺少')
+                        self.label_info.setText('请先输入或选择合适的合约名再点添加')
                 else:
-                    self.label_info.setText('请先输入回测的合约名,该项参数不可缺少')
+                    self.label_info.setText('请先选择合约类型')
             else:
-                self.label_info.setText('没有可用的策略模板,请先添加策略模版')
+                self.label_info.setText('没有可有用策略模板')
         else:
-            self.label_info.setText('没有可用的天勤账户,请先添加可用的天勤账户')
+            self.label_info.setText('请先添加至少一个可用的天勤帐户')
+
 
 
     def start_backtest_process(self, tuple):    # 启动回测进程
@@ -230,14 +259,20 @@ class BackTestWindow(QWidget, UI):  # 创建回测窗口类
         t.start()  # 启动进程
         self.parent.Process_dict[tuple[1]['process_name']] = t.pid
         self.pid = t.pid
+        self.backtest_name = tuple[1]['process_name']
         self.label_info.setText('已成功添加回测进程： ' + tuple[1]['process_name'] + '     进程pid为:  ' + str(t.pid))
 
     def stop_backtest_process(self):
         if self.pid:
-            self.parent.kill_process(self.pid)
-            self.label_info.setText('pid为:  ' + str(self.pid) + '  的回测进程： ' + tuple[1]['process_name'] + '     已关闭')
+            list = self.parent.get_alive_process_pid_list()
+            if self.pid in list:
+                self.parent.kill_process(self.pid)
+                self.label_info.setText('pid为:  ' + str(self.pid) + '  的回测进程： ' + self.backtest_name + '     已关闭')
+                self.pid = None
+            else:
+                self.label_info.setText('  没有运行的 回测进程  ')
         else:
-            self.label_info.setText('回测进程  没有运行  ' )
+            self.label_info.setText('  没有运行的 回测进程  ' )
 
     def get_backtest_date(self):    # 获取回测日期
         start_date = self.calendar_backtest_startdate.selectedDate()
@@ -249,8 +284,9 @@ class BackTestWindow(QWidget, UI):  # 创建回测窗口类
     def process_parameters_input_clear(self):  # 清空回测进程参数输入框
         self.add_paramer_to_combobox()
 
-        self.comboBox_exchange.setCurrentIndex(0)  # 默认交易所      
-        self.symbol.clear()
+        self.comboBox_exchange.setCurrentIndex(0)  # 默认交易所
+        self.comboBox_contract_type.setCurrentIndex(-1)
+        self.comboBox_symbol.clear()
         self.symbol_period.clear()
         self.initial_capital.clear()
         self.web_port.clear()
@@ -267,11 +303,15 @@ class BackTestWindow(QWidget, UI):  # 创建回测窗口类
         self.Customized_parameters6.clear()
         self.Customized_parameters7.clear()
         self.Customized_parameters8.clear()
+        self.Customized_parameters9.clear()
         self.label_backtest_startdate.clear()
         self.label_backtest_enddate.clear()
-        # now = datetime.now()
-        # now_str = now.strftime('%Y.%m.%d-%H时%M分') 
-        # self.label_info.setText(now_str)
+
+        self.checkBox_whether_open_line.setChecked(False)
+        self.checkBox_whether_close_line.setChecked(False)
+        self.open_line_Coordinates.clear()
+        self.close_line_Coordinates.clear()
+
         self.label_info.setText('所有参数已清空，请重新输入')
